Change all int variables to double in OvertimeDTO

+++ src/main/java/com/bionic/dto/OvertimeDTO.java
@@ -3,21 +3,20 @@ package com.bionic.dto;
 import java.util.Date;
 
 public class OvertimeDTO {
-
     private int weekOfYear;
     private Date startTime;
     private Date endTime;
-    private int total;
-    private int paid100;
-    private int paid130;
-    private int paid150;
-    private int paid200;
-    private int waitingdayHours;
-    private int sickdayHours;
-    private int holidayHours;
-    private int atvHours;
-    private int paidLeaveHours;
-    private int unpaidLeaveHours;
+    private double total;
+    private double paid100;
+    private double paid130;
+    private double paid150;
+    private double paid200;
+    private double waitingdayHours;
+    private double sickdayHours;
+    private double holidayHours;
+    private double atvHours;
+    private double paidLeaveHours;
+    private double unpaidLeaveHours;
 
     public int getWeekOfYear() {
         return weekOfYear;
@@ -43,93 +42,92 @@ public class OvertimeDTO {
         this.endTime = endTime;
     }
 
-    public int getTotal() {
+    public double getTotal() {
         return total;
     }
 
-    public void setTotal(int total) {
+    public void setTotal(double total) {
         this.total = total;
     }
 
-    public int getPaid100() {
+    public double getPaid100() {
         return paid100;
     }
 
-    public void setPaid100(int paid100) {
+    public void setPaid100(double paid100) {
         this.paid100 = paid100;
     }
 
-    public int getPaid130() {
+    public double getPaid130() {
         return paid130;
     }
 
-    public void setPaid130(int paid130) {
+    public void setPaid130(double paid130) {
         this.paid130 = paid130;
     }
 
-    public int getPaid150() {
+    public double getPaid150() {
         return paid150;
     }
 
-    public void setPaid150(int paid150) {
+    public void setPaid150(double paid150) {
         this.paid150 = paid150;
     }
 
-    public int getPaid200() {
+    public double getPaid200() {
         return paid200;
     }
 
-    public void setPaid200(int paid200) {
+    public void setPaid200(double paid200) {
         this.paid200 = paid200;
     }
 
-    public int getWaitingdayHours() {
+    public double getWaitingdayHours() {
         return waitingdayHours;
     }
 
-    public void setWaitingdayHours(int waitingdayHours) {
+    public void setWaitingdayHours(double waitingdayHours) {
         this.waitingdayHours = waitingdayHours;
     }
 
-    public int getSickdayHours() {
+    public double getSickdayHours() {
         return sickdayHours;
     }
 
-    public void setSickdayHours(int sickdayHours) {
+    public void setSickdayHours(double sickdayHours) {
         this.sickdayHours = sickdayHours;
     }
 
-    public int getHolidayHours() {
+    public double getHolidayHours() {
         return holidayHours;
     }
 
-    public void setHolidayHours(int holidayHours) {
+    public void setHolidayHours(double holidayHours) {
         this.holidayHours = holidayHours;
     }
 
-    public int getAtvHours() {
+    public double getAtvHours() {
         return atvHours;
     }
 
-    public void setAtvHours(int atvHours) {
+    public void setAtvHours(double atvHours) {
         this.atvHours = atvHours;
     }
 
-    public int getPaidLeaveHours() {
+    public double getPaidLeaveHours() {
         return paidLeaveHours;
     }
 
-    public void setPaidLeaveHours(int paidLeaveHours) {
+    public void setPaidLeaveHours(double paidLeaveHours) {
         this.paidLeaveHours = paidLeaveHours;
     }
 
-    public int getUnpaidLeaveHours() {
+    public double getUnpaidLeaveHours() {
         return unpaidLeaveHours;
     }
 
-    public void setUnpaidLeaveHours(int unpaidLeaveHours) {
+    public void setUnpaidLeaveHours(double unpaidLeaveHours) {
         this.unpaidLeaveHours = unpaidLeaveHours;
     }
-
-
 }
+
